Corrects reaction prediction for lava burst
Current Behavior:

When lava burst is being checked, the cast prediction does not correctly predict any further filler spells because it is still available. If you are casting lava burst without a lava surge proc, it will predict that you should be casting lava burst next, even if it will go on cooldown.

The changes look for both cases:

1. if you aren't casting lava burst, but should
2. if you are casting lava burst and you have a lava surge proc waiting

To replicate: Cast lava burst without a proc, and it will show lava burst when it should show the next filler spell (LB/CL) or Earth shock if you are going to hit maelstrom.