fix:(blood) nf rp gen and cap heart strike hits
The current APL assumes you have the covenant legendary equipped and the *1.2 rp gen for Night Fae was placed in the wrong section of the if.

EnemiesMeleeCount was changed to HeartStrikeCount which is the EnemiesMeleeCount clamped to 5. Previous logic would allow for incorrect runic power generation during aoe with Heartbreaker.

I'm not sure how the Unity legendary is going to work so I added into here for now. The logic might be better suited for HeroLib if we have to query all the Unity legendary IDs.